Fix tests and licensing
