+++ minecraft-test/src/main/java/org/quiltmc/test/EntrypointTest.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2016 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.test;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public final class EntrypointTest {
+	private static final Logger LOGGER = LogManager.getLogger();
+	public static final CustomEntry FIELD_ENTRY = EntrypointTest::fieldEntry;
+
+	public static String staticEntry() {
+		return "static";
+	}
+
+	public EntrypointTest() {
+		LOGGER.info("EntrypointTest instance created");
+	}
+
+	public String instanceEntry() {
+		return "instance";
+	}
+
+	public static String fieldEntry() {
+		return "field";
+	}
+}
