--- src/test/java/net/fabricmc/test/CustomEntry.java
@@ -1,9 +1,0 @@
-package net.fabricmc.test;
-
-@FunctionalInterface
-public interface CustomEntry {
-	String name();
-	default String describe() {
-		return "Custom entry point \"" + name() + '"';
-	}
-}
